refactor: shortcuts for slide and elements
