+++ frontend/src/components/ImageElement.vue
@@ -4,7 +4,7 @@
 
 <script setup>
 import { computed } from 'vue'
-import { setActiveElement } from '@/stores/slide'
+import { setActiveElement } from '@/stores/element'
 
 const element = defineModel('element', {
 	type: Object,
