+++ frontend/src/components/TextPropertyTab.vue
@@ -130,7 +130,7 @@ import { FormControl } from 'frappe-ui'
 import { Strikethrough, CaseUpper } from 'lucide-vue-next'
 
 import { debounce } from '@/utils/debounce'
-import { activeElement } from '@/stores/slide'
+import { activeElement } from '@/stores/element'
 import SliderInput from './SliderInput.vue'
 import NumberInput from './NumberInput.vue'
 import ColorPicker from './ColorPicker.vue'
