+++ frontend/src/components/VideoElement.vue
@@ -22,7 +22,8 @@
 
 <script setup>
 import { ref, useTemplateRef, computed, useAttrs } from 'vue'
-import { activeElement, currentDataIndex, inSlideShow, setActiveElement } from '@/stores/slide'
+import { inSlideShow } from '@/stores/slide'
+import { currentDataIndex, setActiveElement } from '@/stores/element'
 
 const attrs = useAttrs()
